Added type type conversions and comment
unnecessary type conversions to uint64 is needed for FreeBSD

+++ guerrilla_unix.go
@@ -11,5 +11,6 @@ func getFileLimit() (uint64, error) {
 	if err != nil {
 		return 0, err
 	}
-	return rLimit.Max, nil
+	//unnecessary type conversions to uint64 is needed for FreeBSD
+	return uint64(rLimit.Max), nil
 }
